rename: path 변경 적용
